feat(mcp): add core tools and types for project management
- Implemented tool functions for inspecting, validating, creating services, adding resources, defining resource schemas, adding API versions, generating code, syncing dependencies, building, testing, and performing smoke tests on projects.
- Introduced mcpServer struct to manage workspace boundaries and tool calls.
- Defined JSON-RPC message structures for requests and responses.
- Added resource schema types and error handling for tool execution.
