+++ cmd/fabrica/config_security_test.go
@@ -7,44 +7,46 @@ package main
 import (
 	"bytes"
 	"testing"
+
+	"github.com/openchami/fabrica/internal/config"
 )
 
 func TestValidateConfig_AuthZRequiresAuthN(t *testing.T) {
-	cfg := NewDefaultConfig("test", "example.com/test")
+	cfg := config.NewDefaultConfig("test", "example.com/test")
 	cfg.Features.Security.AuthZ.Enabled = true
 	cfg.Features.Security.AuthN.Enabled = false
 
-	err := ValidateConfig(cfg)
+	err := config.ValidateConfig(cfg)
 	if err == nil {
 		t.Fatalf("expected error")
 	}
 }
 
 func TestValidateConfig_RequiresStorage(t *testing.T) {
-	cfg := NewDefaultConfig("test", "example.com/test")
+	cfg := config.NewDefaultConfig("test", "example.com/test")
 	cfg.Features.Storage.Enabled = false
 
-	if err := ValidateConfig(cfg); err == nil {
+	if err := config.ValidateConfig(cfg); err == nil {
 		t.Fatalf("expected storage-disabled config to be rejected")
 	}
 }
 
 func TestValidateConfig_EventsBusMemoryOnly(t *testing.T) {
-	cfg := NewDefaultConfig("test", "example.com/test")
+	cfg := config.NewDefaultConfig("test", "example.com/test")
 	cfg.Features.Events.Enabled = true
 	cfg.Features.Events.BusType = "nats"
 
-	if err := ValidateConfig(cfg); err == nil {
+	if err := config.ValidateConfig(cfg); err == nil {
 		t.Fatalf("expected non-memory event bus to be rejected")
 	}
 }
 
 func TestValidateConfig_ReconciliationRequiresEvents(t *testing.T) {
-	cfg := NewDefaultConfig("test", "example.com/test")
+	cfg := config.NewDefaultConfig("test", "example.com/test")
 	cfg.Features.Reconciliation.Enabled = true
 	cfg.Features.Events.Enabled = false
 
-	if err := ValidateConfig(cfg); err == nil {
+	if err := config.ValidateConfig(cfg); err == nil {
 		t.Fatalf("expected reconciliation without events to be rejected")
 	}
 }
@@ -52,9 +54,9 @@ func TestValidateConfig_ReconciliationRequiresEvents(t *testing.T) {
 func TestValidateConfig_InvalidAuthZModeDefaultsToEnforceAndWarns(t *testing.T) {
 	var warn bytes.Buffer
 
-	mode := normalizeSecurityMode(SecurityMode("bogus"), &warn)
-	if mode != SecurityModeEnforce {
-		t.Fatalf("expected %q, got %q", SecurityModeEnforce, mode)
+	mode := config.NormalizeSecurityMode(config.SecurityMode("bogus"), &warn)
+	if mode != config.SecurityModeEnforce {
+		t.Fatalf("expected %q, got %q", config.SecurityModeEnforce, mode)
 	}
 	if warn.Len() == 0 {
 		t.Fatalf("expected warning output")
@@ -64,9 +66,9 @@ func TestValidateConfig_InvalidAuthZModeDefaultsToEnforceAndWarns(t *testing.T) 
 func TestValidateConfig_EmptyAuthZModeDefaultsToEnforceWithoutWarning(t *testing.T) {
 	var warn bytes.Buffer
 
-	mode := normalizeSecurityMode("", &warn)
-	if mode != SecurityModeEnforce {
-		t.Fatalf("expected %q, got %q", SecurityModeEnforce, mode)
+	mode := config.NormalizeSecurityMode("", &warn)
+	if mode != config.SecurityModeEnforce {
+		t.Fatalf("expected %q, got %q", config.SecurityModeEnforce, mode)
 	}
 	if warn.Len() != 0 {
 		t.Fatalf("expected no warning output")
@@ -90,7 +92,7 @@ func TestCreateFabricaConfig_WithAuthEnablesSecurityAuthN(t *testing.T) {
 		t.Fatalf("createFabricaConfig returned error: %v", err)
 	}
 
-	cfg, err := LoadConfig(dir)
+	cfg, err := config.LoadConfig(dir)
 	if err != nil {
 		t.Fatalf("LoadConfig returned error: %v", err)
 	}
@@ -104,7 +106,7 @@ func TestCreateFabricaConfig_WithAuthEnablesSecurityAuthN(t *testing.T) {
 	if cfg.Features.Security.AuthZ.Enabled {
 		t.Fatalf("expected security.authz.enabled to remain disabled")
 	}
-	if cfg.Features.Security.AuthZ.Mode != SecurityModeEnforce {
-		t.Fatalf("expected default authz mode %q, got %q", SecurityModeEnforce, cfg.Features.Security.AuthZ.Mode)
+	if cfg.Features.Security.AuthZ.Mode != config.SecurityModeEnforce {
+		t.Fatalf("expected default authz mode %q, got %q", config.SecurityModeEnforce, cfg.Features.Security.AuthZ.Mode)
 	}
 }
